Moved Armature Validation to it's own file
